requirements.txt
openai>=1.0.0
ddgs>=0.1.0
wikipedia-api>=0.6.0
gradio>=4.0.0
python-dotenv>=1.0.0
